Separate classifier from HMM model
