+++ lib/model/hmm.js
@@ -461,59 +461,3 @@ export class ContinuousHMM extends HMMBase {
 		return v
 	}
 }
-
-/**
- * Hidden Markov model classifier
- */
-export class HMMClassifier {
-	/**
-	 * @param {number} n Number of states
-	 * @param {*} cls HMM class
-	 */
-	constructor(n, cls = ContinuousHMM) {
-		this._n = n
-		this._classes = null
-		this._models = []
-		this._modelcls = cls
-	}
-
-	/**
-	 * Fit model.
-	 *
-	 * @param {Array<Array<number>> | Array<Array<Array<number>>>} x Training data
-	 * @param {*[]} y Target values
-	 * @param {boolean} scaled Do scaled calculation or not
-	 */
-	fit(x, y, scaled = false) {
-		if (!this._classes) {
-			this._classes = [...new Set(y)]
-			for (let i = 0; i < this._classes.length; i++) {
-				const m = new this._modelcls(this._n)
-				this._models.push(m)
-			}
-		}
-		for (let i = 0; i < this._models.length; i++) {
-			const xi = x.filter((v, j) => this._classes[i] === y[j])
-			this._models[i].fit(xi, scaled)
-		}
-	}
-
-	/**
-	 * Returns predicted values.
-	 *
-	 * @param {Array<Array<number>> | Array<Array<Array<number>>>} x Sample data
-	 * @returns {(* | null)[]} Predicted values
-	 */
-	predict(x) {
-		const pred = []
-		for (const model of this._models) {
-			pred.push(model.probability(x))
-		}
-
-		const pm = Matrix.fromArray(pred)
-		const p = pm.argmax(0)
-		const ps = pm.max(0)
-		p.map((v, i) => (ps.at(i) > 0 ? this._classes[v] : null))
-		return p.value
-	}
-}
